Mdm schema migration (#1311)
* fix schema endpoint mdm tests

* fix mdm testMigrateByESSchemaNull

* make sure changes in versionEdge migration status are persisted

* Fix issues with admin permissions change withing an existing transaction

* Refactor crud handlers to avoid using multiple tx callbacks

* Refactor binary upload handlers to avoid using multiple tx callbacks

* Fix field migrations

* Fix reference updates events

* Fix project version purge

* test fixes for mdm

* test fixes for mdm

* remove unneded merge into persisted calls

* javadoc

* Remove nested transaction (makes version purge job execute immediately
when run with hibernate)

+++ tests/tests-core/src/main/java/com/gentics/mesh/core/schema/field/StringFieldMigrationTest.java
@@ -285,12 +285,12 @@ public class StringFieldMigrationTest extends AbstractFieldMigrationTest impleme
 	public void testChangeToNumber() throws Exception {
 		changeType(CREATESTRING, FILL0, FETCH, CREATENUMBER, (container, name) -> {
 			assertThat(container.getNumber(name)).as(NEWFIELD).isNotNull();
-			assertThat(container.getNumber(name).getNumber()).as(NEWFIELDVALUE).isEqualTo(0L);
+			assertThat(container.getNumber(name).getNumber().longValue()).as(NEWFIELDVALUE).isEqualTo(0L);
 		});
 
 		changeType(CREATESTRING, FILL1, FETCH, CREATENUMBER, (container, name) -> {
 			assertThat(container.getNumber(name)).as(NEWFIELD).isNotNull();
-			assertThat(container.getNumber(name).getNumber()).as(NEWFIELDVALUE).isEqualTo(1L);
+			assertThat(container.getNumber(name).getNumber().longValue()).as(NEWFIELDVALUE).isEqualTo(1L);
 		});
 
 		changeType(CREATESTRING, FILLTEXT, FETCH, CREATENUMBER, (container, name) -> {
